Move kubernetes config parsing to util
